requirements_flexible.txt
# Flexible requirements - let pip resolve versions
# Core dependencies
flask>=2.3.0
flask-cors>=4.0.0
pandas>=2.0.0
numpy>=1.23.0,<1.26.0,!=1.24.0  # Avoid the problematic 1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
redis>=4.6.0
plotly>=5.17.0
openpyxl>=3.1.2
pyyaml>=6.0

# NLP and embeddings
nltk>=3.8
spacy>=3.6.0
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0
faiss-cpu>=1.7.4

# PDF processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0

# LLM APIs
openai>=1.0.0
anthropic>=0.7.0

# Development and testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
ipython>=8.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# API documentation
flasgger>=0.9.7